fix: replace file function & documentation (how to running the app)
